+++ namer/consul/src/main/scala/io/buoyant/namer/consul/ConsulInitializer.scala
@@ -43,9 +43,6 @@ case class ConsulConfig(
 ) extends NamerConfig {
 
   @JsonIgnore
-  override val experimentalRequired = true
-
-  @JsonIgnore
   override def defaultPrefix: Path = Path.read("/io.l5d.consul")
 
   private[this] def getHost = host.getOrElse("localhost")