+++ CHANGES.md
@@ -1,6 +1,9 @@
 ## x.x.x
 
 * Change default value of `attemptTTwitterUpgrade` to `false`
+* The `io.l5d.consul` and `io.l5d.k8s` namers are no longer experimental
+* Fix the `h2` protocol to gracefully handle connection loss and
+  stream interruption.
 * Add `io.l5d.static` identifier
 
 ## 0.8.3